rename to stream
