+++ compiler/accelerators/snax.py
@@ -12,7 +12,7 @@ from compiler.accelerators.streamers import StreamerConfiguration
 from compiler.accelerators.streamers.streamers import StreamerFlag, StreamerOpts
 from compiler.dialects import accfg, stream
 from compiler.dialects.snax_stream import StreamerConfigurationAttr, StreamingRegionOp
-from compiler.ir.flow import Template
+from compiler.ir.stream import Template
 
 c0_attr = builtin.IntegerAttr(0, builtin.IndexType())
