+++ compiler/ir/stream/access_pattern.py
